declare events in the modules themseleves
