+++ examples/action_events/mod.js
@@ -9,8 +9,8 @@ let span = document.querySelector("span");
 document.addEventListener("#action", function (e) {
     if (!span)
         return;
-    let { kind } = e.action;
-    if ("increase_count" === kind) {
+    let { type } = e.action;
+    if ("increase_count" === type) {
         count += 1;
         span.textContent = count.toString();
     }
