+++ examples/throttle/mod.js
@@ -6,7 +6,7 @@ const _hyperEvents = new HyperEvents({
 });
 const ul = document.querySelector("ul");
 document.addEventListener("#action", function (e) {
-    let { timeStamp, type } = e.action.originEvent;
+    let { timeStamp, type } = e.action.event;
     const li = document.createElement("li");
     li.textContent = `action occured at <${timeStamp.toFixed(2)}> on <${type}>`;
     ul?.append(li);
